adjust tests to play with pytest.bdd
