+++ tests/conftest.py
@@ -1,0 +1,54 @@
+import os
+import pytest
+from sqlalchemy import create_engine
+
+import journal
+
+TEST_DATABASE_URL = os.environ.get(
+    'DATABASE_URL',
+    'postgresql://ajw@localhost:5432/test-learning-journal'
+)
+os.environ['DATABASE_URL'] = TEST_DATABASE_URL
+os.environ['TESTING'] = "True"
+
+
+@pytest.fixture(scope='session')
+def connection(request):
+    engine = create_engine(TEST_DATABASE_URL)
+    journal.Base.metadata.create_all(engine)
+    connection = engine.connect()
+    journal.DBSession.registry.clear()
+    journal.DBSession.configure(bind=connection)
+    journal.Base.metadata.bind = engine
+    request.addfinalizer(journal.Base.metadata.drop_all)
+    return connection
+
+
+@pytest.fixture()
+def db_session(request, connection):
+    from transaction import abort
+    trans = connection.begin()
+    request.addfinalizer(trans.rollback)
+    request.addfinalizer(abort)
+
+    from journal import DBSession
+    return DBSession
+
+
+@pytest.fixture()
+def app():
+    from journal import main
+    from webtest import TestApp
+    app = main()
+    return TestApp(app)
+
+
+@pytest.fixture()
+def entry(db_session):
+    entry = journal.Entry.write(
+        title='Test Title',
+        text='Test Entry Text',
+        session=db_session
+    )
+    db_session.flush()
+    return entry
